requirements.txt
djangorestframework
psycopg2-binary
corsheaders